Pull bash configuration into a series of functions
PS1 and competion are big enough for seperate functions, while the
various shopt settings go in the main function. I pondered a lot
whether or not to pull the bash configuration into a seperat file
as bashrc /is/ bash's configuration/setup. In the end I did it primarily
for the aesthetics of bashrc.