Change PS1 to show the time and take 2 lines
It's closer to the prompt I recently changed for virtualenvs, but with
the time on the right hand side. The time is made to always be on the right
hand side using tput to query the current size. There's a lot of jankiness
with how the colors are included though.