Rewrite PS1 configurator, add environment support
The comments are better organized, tputs has been replaced by $COLUMNS, a
builtin Bash variable that is automatically updated. The ability to add an
optional environment is also added. The intended use for this is by RVM and
virtualenvwrapper, but could easily be used for jails and other
environment-y things.