Rewrite the dotfiles linking step
The first advantage is that this is in its own function now. The second
is that it's conceptually cleaner, only linking files while creating
directories. When updating the links, it only updates old links, and
only delete empty directories. This should allow site-specific
modifications to be made without getting them blown away when the links
are updated. It's also more effcient in creating/removing links, only
creating links that need to be updated, and only removing broken links.